+++ examples/resources/sonatyperepo_repository_alpine_proxy/import.sh
@@ -1,0 +1,6 @@
+# Existing proxy alpine repository configuration can be imported as follows.
+#
+# NOTE: The Identifier REPOSITORY_NAME needs to match repository name in your sonatype nexus repository instance.
+
+# Example
+terraform import sonatyperepo_repository_alpine_proxy.example REPOSITORY_NAME